service: Init a PG database
When running locally I'd rather use PG as that's the database system on
production systems too. As such I started using PG locally.

For some local things you need to understand what PG does, and how. To
ease this onboarding to PG for others I figured I'd make it an optional
component for this project.

This change hooks into the existing `init` command and sets up PG iff
the users has set `postgres` in their flowforge package. For a user to
use this feature this implies they'll have to run `init` twice when
using PG.