init: Command is idempotent when using PG
`npm run init` has been idempotent, but that behaviour changed since
HEAD~ when using a PG database. `initdb` exits with a non-zero exit code
when the data directory isn't empty as there might be valuable data in
that directory about to be overriden.

Which this change the same test is performed as `initdb` does, are there
any file in the directory. If not, initiate the database. Else, we don't
execute `initdb`. This in turn make `npm run init` idempotent again.